Fixes #69
The object mask used for public and private image templates is different.  This change accounts for that difference so that the routine SoftLayer::ImageTemplate.template_with_global_id returns the correct template.

Note that the routine now throws an exception if a request for a given template (by global id) does not return a unique template.